+++ src/main/java/org/mesdag/particlestorm/data/molang/compiler/function/round/TruncateFunction.java
@@ -4,14 +4,11 @@ import org.mesdag.particlestorm.api.MolangInstance;
 import org.mesdag.particlestorm.data.molang.compiler.MathValue;
 import org.mesdag.particlestorm.data.molang.compiler.function.MathFunction;
 
-/**
- * {@link MathFunction} value supplier
- *
- * <p>
- * <b>Contract:</b>
- * <br>
- * Returns the closest value that is equal to the input value or closer to zero, and is equal to an integer
- */
+/// [MathFunction] value supplier
+///
+/// **Contract:**
+///
+/// Returns the closest value that is equal to the input value or closer to zero, and is equal to an integer
 public final class TruncateFunction extends MathFunction {
     private final MathValue value;
 
@@ -27,8 +24,8 @@ public final class TruncateFunction extends MathFunction {
     }
 
     @Override
-    public double compute(MolangInstance instance) {
-        return (long)this.value.get(instance);
+    public float compute(MolangInstance instance) {
+        return (int) this.value.get(instance);
     }
 
     @Override
@@ -38,6 +35,6 @@ public final class TruncateFunction extends MathFunction {
 
     @Override
     public MathValue[] getArgs() {
-        return new MathValue[] {this.value};
+        return new MathValue[]{this.value};
     }
 }
