+++ src/main/java/org/mesdag/particlestorm/mixin/EntityMixin.java
@@ -22,9 +22,9 @@ public abstract class EntityMixin implements IEntity {
             Hashtable<String, Variable> table = new Hashtable<>();
             table.put("variable.entity_scale", new Variable("variable.entity_scale", p -> {
                 if (p.getAttachedEntity() instanceof LivingEntity living) {
-                    return living.getAttributeValue(Attributes.SCALE);
+                    return (float) living.getAttributeValue(Attributes.SCALE);
                 }
-                return 1.0;
+                return 1;
             }));
             this.particlestorm$variableTable = new VariableTable(table, null);
         }
